Fix test case and refactor
